requirements.txt
asgiref==3.5.1
bcrypt==3.2.2
boto3==1.22.7
botocore==1.25.7
cffi==1.15.0
cryptography==37.0.2
Django==4.0.4
django-cors-headers==3.11.0
django-storages==1.12.3
djangorestframework==3.13.1
jmespath==1.0.0
paramiko==2.10.4
psycopg2-binary==2.9.3
pycparser==2.21
PyNaCl==1.5.0
python-dateutil==2.8.2
pytz==2022.1
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
storage==0.0.4.3
urllib3==1.26.9
